# ENH: remove unnecessary root-find from skewnorm