<!-- 
Thanks for contributing a pull request! Please ensure that
your PR satisfies the checklist before submitting:
http://scipy.github.io/devdocs/dev/contributor/development_workflow.html#checklist-before-submitting-a-pr

Also, please name and describe your PR as you would write a
commit message:
http://scipy.github.io/devdocs/dev/contributor/development_workflow.html#writing-the-commit-message

Depending on your changes, you can skip CI operations and save time and energy: 
http://scipy.github.io/devdocs/dev/contributor/continuous_integration.html#skipping

Note that we are a team of volunteers; we appreciate your
patience during the review process.

Again, thanks for contributing!
-->

#### Reference issue
None

#### What does this implement/fix?
Previously the `skew_d` function was inverted using `root_scalar`. However, there is a simple closed-form inverse, so use that instead.

#### Additional information
Reference for inverse function:
<https://en.wikipedia.org/wiki/Skew_normal_distribution#Estimation>

To demonstrate that the inverse function is correct:
```python
np.array_equal(d_skew([-s_max,s_max]), [-1, 1])
np.array_equal(d_skew(np.array([-s_max,s_max])), [-1, 1])

d_vals = np.linspace(-1, 1, 100)
s_vals = np.linspace(-s_max, s_max, 100)
assert_almost_equal(d_vals, d_skew(skew_d(d_vals)), decimal=15)
assert_almost_equal(s_vals, skew_d(d_skew(s_vals)), decimal=15)
```
I'm not sure how I could currently turn these into unit tests as the functions are only defined within the `skewnorm.fit` scope. If desired I could refactor them to be top-level in `scipy.stats._continuous_distns` and add these as tests.

The performance improvement is not as dramatic as I expected:
```python
from scipy.stats import skewnorm

# before
%timeit skewnorm.fit([1,2,3,3,4,4,4,5], method="mm")
673 µs ± 18.3 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)

# after
%timeit skewnorm.fit([1,2,3,3,4,4,4,5], method="mm")
616 µs ± 10.1 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```
This is because the most expensive call is `s = stats.skew(data)`, while `root_scalar` is actually quite fast.