@mdhaber 2023-07-02
Please confirm that this test would fail if the formula for the increase were wrong.
https://github.com/scipy/scipy/blob/cfe80116aaa145061246b8aec0e98248fbefb835/scipy/stats/tests/test_distributions.py#L3625-L3633

If so, you don't need a new test. If not, write a test along those lines.