@maresb 2023-07-02
Ah, good call. I added a small error of `1e-7` to the result of `d_skew` and the last line with `assert_allclose` does indeed fail.

As suggested I just added a few checks to make sure fits with MoM and MLE behave as expected under sign change.